chore(docker-compose): split docker-compose and move to own submodule
