+++ src/docker-compose/index.ts
@@ -1,0 +1,4 @@
+export * from "./docker-compose-port";
+export * from "./docker-compose-service";
+export * from "./docker-compose-volume";
+export * from "./docker-compose";
